Recursive Array Functions

These are functions designed for recursive arrays, like arrays of arrays, and do not require that the RecursiveArrayTools types are used.

Function List

RecursiveArrayTools.recursivecopy — Function

recursivecopy(a::Union{AbstractArray{T, N}, AbstractVectorOfArray{T, N}})

A recursive copy function. Acts like a deepcopy on arrays of arrays, but like copy on arrays of scalars.

RecursiveArrayTools.recursivecopy! — Function

recursivecopy!(b::AbstractArray{T, N}, a::AbstractArray{T, N})

A recursive copy! function. Acts like a deepcopy! on arrays of arrays, but like copy! on arrays of scalars.

RecursiveArrayTools.vecvecapply — Function

vecvecapply(f::Base.Callable, v)

Calls f on each element of a vecvec v.

RecursiveArrayTools.copyat_or_push! — Function

copyat_or_push!{T}(a::AbstractVector{T}, i::Int, x)

If i<length(x), it's simply a recursivecopy! to the ith element. Otherwise, it will push! a deepcopy.